ci: Automate the SBOM generation for all projects with fastlane
IPC-68
